backend/requirements.txt
python-dotenv==1.1.1
httpx==0.28.1
fastapi==0.115.6
uvicorn[standard]==0.34.0
pyjwt==2.10.1

